+++ src/App/Documentation/components/Buttons/index.js
@@ -4,12 +4,13 @@ import PrismCode from "react-prism";
 
 import { ComponentPreview, Attribute, Property, DocToc } from "#";
 import ButtonComponent from "@/Button";
+import Alert from "@/Alert";
 import { button } from "$/px-script";
 
 const Examples = () => (
     <>
         <h2 id="examples">Examples</h2>
-        <p>Several button styles etc...</p>
+        <p>The DesignGuide includes a few predefined button styles, each serving its own semantic purpose.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent type="primary" label="Primary" />{"\n"}
@@ -24,7 +25,8 @@ const Examples = () => (
 const UsageWithOtherTags = () => (
     <>
         <h2 id="usage-with-other-tags">Usage with other tags</h2>
-        <p>The <Property value=".btn" /> class can also be used with other html-tags like <PrismCode className="language-html">{"<a>"}</PrismCode> or <PrismCode className="language-html">{"<input>"}</PrismCode>.</p>
+        <p>The <Property value=".btn" /> classes are designed to be used with the <PrismCode className="language-html">{"<button>"}</PrismCode> element. However, you can also use these classes on <PrismCode className="language-html">{"<a>"}</PrismCode> or <PrismCode className="language-html">{"<input>"}</PrismCode> elements (though some browsers may apply a slightly different rendering).</p>
+        <p>When using button classes on <PrismCode className="language-html">{"<a>"}</PrismCode> elements that are used to trigger in-page functionality (like collapsing content), rather than linking to new pages or sections within the current page, these links should be given a <Attribute name="role" value="button" /> to appropriately convey their purpose to assistive technologies such as screen readers.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent type="primary" label="Anchor" href="#" />{"\n"}
@@ -40,7 +42,7 @@ const UsageWithOtherTags = () => (
 const OutlineButtons = () => (
     <>
         <h2 id="outline-buttons">Outline buttons</h2>
-        <p>Several button styles etc...<Property value=".btn-outline-{variant}" /></p>
+        <p>Replace the default modifier classes with the <Property value=".btn-outline-{style}" /> ones to get an outline of the button style.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent outline type="primary" label="Primary" />{"\n"}
@@ -68,25 +70,22 @@ const ButtonGroup = () => (
 const Sizes = () => (
     <>
         <h2 id="sizes">Sizes</h2>
-        <p>Several button sizes etc...</p>
-        <h3>Large buttons</h3>
-        <p>...<Property value=".btn-lg" /></p>
+        <p>For a different sized button add <Property value=".btn-lg" />, <Property value=".btn-sm" /> or <Property value=".btn-xs" />.</p>
+        <h3>Large</h3>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent size="lg" type="primary" label="Large button" />{"\n"}
                 <ButtonComponent size="lg" type="secondary" label="Large button" />{"\n"}
             </div>
         </ComponentPreview>
-        <h3>Small buttons</h3>
-        <p>...<Property value=".btn-sm" /></p>
+        <h3>Small</h3>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent size="sm" type="primary" label="Small button" />{"\n"}
                 <ButtonComponent size="sm" type="secondary" label="Small button" />{"\n"}
             </div>
         </ComponentPreview>
-        <h3>Extra small buttons</h3>
-        <p>...<Property value=".btn-xs" /></p>
+        <h3>Extra small</h3>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent size="xs" type="primary" label="Extra small button" />{"\n"}
@@ -94,7 +93,7 @@ const Sizes = () => (
             </div>
         </ComponentPreview>
         <h3>Block level</h3>
-        <p>...<Property value=".btn-block" /></p>
+        <p>Create block level buttons—those that span the full width of a parent by adding <Property value=".btn-block" />.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent fullWidth type="primary" label="Block level button" />{"\n"}
@@ -107,12 +106,12 @@ const Sizes = () => (
 const ActiveState = () => (
     <>
         <h2 id="active-state">Active state</h2>
-        <p>Active state with <Property value=".active" />...</p>
+        <p>Buttons will appear pressed (with a darker background, darker border, and inset shadow) when active. <b>There’s no need to add a class to <PrismCode className="language-html">{"<button>"}</PrismCode>s as they use a pseudo-class.</b> However, you can still force the same active appearance with <Property value=".active" /> (and include the <Attribute name="aria-pressed" value="true" /> attribute) should you need to replicate the state programmatically.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
-                <ButtonComponent active type="primary" label="Primary" />{"\n"}
-                <ButtonComponent active type="secondary" label="Secondary" />{"\n"}
-                <ButtonComponent active type="danger" label="Danger" />{"\n"}
+                <ButtonComponent href="#" active type="primary" label="Primary" />{"\n"}
+                <ButtonComponent href="#" active type="secondary" label="Secondary" />{"\n"}
+                <ButtonComponent href="#" active type="danger" label="Danger" />{"\n"}
             </div>
         </ComponentPreview>
     </>
@@ -121,7 +120,7 @@ const ActiveState = () => (
 const DisabledState = () => (
     <>
         <h2 id="disabled-state">Disabled state</h2>
-        <p>Disabled state with <Attribute name="disabled" />...</p>
+        <p>Make buttons look inactive by adding the <Attribute name="disabled" /> boolean attribute to any <PrismCode className="language-html">{"<button>"}</PrismCode> element.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent disabled type="primary" label="Primary" />{"\n"}
@@ -129,17 +128,36 @@ const DisabledState = () => (
                 <ButtonComponent disabled type="danger" label="Danger" />{"\n"}
             </div>
         </ComponentPreview>
+        <p>Disabled buttons using the <PrismCode className="language-html">{"<a>"}</PrismCode> element behave a bit different:</p>
+        <ul>
+            <li><PrismCode className="language-html">{"<a>"}</PrismCode>s do not support the <Attribute name="disabled" />attribute, so you must add the <Property value=".disabled" /> class to make it visually appear disabled.</li>
+            <li>Some future-friendly styles are included to disable all <Property value="pointer-events" /> on anchor buttons. In browsers which support that property, you won’t see the disabled cursor at all.</li>
+            <li>Disabled buttons should include the <Attribute name="aria-disabled" value="true" /> attribute to indicate the state of the element to assistive technologies.</li>
+        </ul>
+        <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
+            <div className="button-group">{"\n"}
+                <ButtonComponent href="#" disabled type="primary" label="Primary" />{"\n"}
+                <ButtonComponent href="#" disabled type="secondary" label="Secondary" />{"\n"}
+                <ButtonComponent href="#" disabled type="danger" label="Danger" />{"\n"}
+            </div>
+        </ComponentPreview>
+        <Alert type="warning">
+            <h5>Link functionality caveat</h5>
+            <p>The <Property value=".disabled" /> class uses <Property value="pointer-events: none" /> to try to disable the link functionality of <PrismCode className="language-html">{"<a>"}</PrismCode>s, but that CSS property is not yet standardized. In addition, even in browsers that do support <Property value="pointer-events: none" />, keyboard navigation remains unaffected, meaning that sighted keyboard users and users of assistive technologies will still be able to activate these links. So to be safe, add a <Property value={"tabindex=\"-1\""} /> attribute on these links (to prevent them from receiving keyboard focus) and use custom JavaScript to disable their functionality.</p>
+        </Alert>
     </>
 );
 
 const UsageWithIcons = () => (
     <>
         <h2 id="usage-with-icons">Usage with icons</h2>
-        <p>To use a button with an icon... Read more about icon usage here <Link to="/docs/core/icons">here</Link>.</p>
+        <p>To use a button with an icon simply put the icon markup inside the <PrismCode className="language-html">{"<button>"}</PrismCode>.</p>
+        <p>Read more about icon usage here <Link to="/docs/core/icons">here</Link>.</p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent icon="cloud" type="primary" label="Primary" />{"\n\n"}
                 <ButtonComponent icon="cloud" type="secondary" label="Secondary" />{"\n"}
+                <ButtonComponent icon="cloud" type="danger" label="Danger" />{"\n"}
             </div>
         </ComponentPreview>
     </>
@@ -156,7 +174,8 @@ const ButtonLoader = () => (
                 <ButtonComponent type="danger" label="Danger" loader loading />{"\n"}
             </div>
         </ComponentPreview>
-        <p>With <Property value="disabled" /></p>
+        <h3>Disabled loader</h3>
+        <p><Property value=".loading" /> with <Attribute name="disabled" /></p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent disabled type="primary" label="Primary" loader loading />{"\n"}
@@ -164,7 +183,8 @@ const ButtonLoader = () => (
                 <ButtonComponent disabled type="danger" label="Danger" loader loading />{"\n"}
             </div>
         </ComponentPreview>
-        <p>With <Property value=".outline" /></p>
+        <h3>Outline loader</h3>
+        <p><Property value=".loading" /> with <Property value=".outline" /></p>
         <ComponentPreview language="html" showCasePanel codeFigure removeOuterTag>
             <div className="button-group">{"\n"}
                 <ButtonComponent outline type="primary" label="Primary" loader loading />{"\n"}
@@ -172,6 +192,7 @@ const ButtonLoader = () => (
                 <ButtonComponent outline type="danger" label="Danger" loader loading />{"\n"}
             </div>
         </ComponentPreview>
+        <h3>The markup</h3>
         <p>Adding the attribute <Attribute data name="button-loader" /> to a button component will produce the following html:</p>
         <ComponentPreview language="html" codeFigure>
             <button className="btn btn-default loading" type="button" data-button-loader>{"\n\t"}
@@ -185,13 +206,16 @@ const ButtonLoader = () => (
                 </div>{"\n"}
             </button>
         </ComponentPreview>
-        <p>If you want more control over the loader component you can include the html yourself (e.g. for server-side rendering), just make sure you <b>don{"'"}t</b> add the attribute <Attribute data name="button-loader" />.</p>
+        <Alert type="warning">
+            <h5>Including the loader html</h5>
+            <p>If you want more control over the loader component you can include the html yourself (e.g. for server-side rendering), just make sure you do <b>not</b> add the attribute <Attribute data name="button-loader" />.</p>
+        </Alert>
     </>
 );
 
 const ButtonsText = () => (
     <div className="col-md-12 col-lg-10 doc-body">
-        <p className="lead">Use buttons...</p>
+        <p className="lead">Use our button styles for actions in forms, dialogs, and more with support for multiple sizes, states, and more.</p>
         <Examples />
         <UsageWithOtherTags />
         <OutlineButtons />
@@ -224,4 +248,4 @@ class Buttons extends Component {
 export default Buttons;
 
 /* for testing */
-export { Examples, UsageWithOtherTags, OutlineButtons, Sizes, ActiveState, DisabledState, UsageWithIcons, ButtonLoader, ButtonsText, ButtonGroup };
+export { Examples, UsageWithOtherTags, OutlineButtons, ButtonGroup, Sizes, ActiveState, DisabledState, UsageWithIcons, ButtonLoader, ButtonsText };