+++ src/App/Documentation/components/Buttons/index.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 
-import Buttons, { Examples, UsageWithOtherTags, OutlineButtons, Sizes, ActiveState, DisabledState, UsageWithIcons, ButtonLoader, ButtonsText } from "./index";
+import Buttons, { Examples, UsageWithOtherTags, OutlineButtons, ButtonGroup, Sizes, ActiveState, DisabledState, UsageWithIcons, ButtonLoader, ButtonsText } from "./index";
 
 describe("Documentation: Buttons", () => {
     it("is defined", () => {
@@ -45,6 +45,18 @@ describe("Documentation: Buttons", () => {
 
         it("renders", () => {
             const wrapper = shallow(<OutlineButtons />);
+
+            expect(wrapper).toMatchSnapshot();
+        });
+    });
+
+    describe("Button group", () => {
+        it("is defined", () => {
+            expect(ButtonGroup).toBeDefined();
+        });
+
+        it("renders", () => {
+            const wrapper = shallow(<ButtonGroup />);
 
             expect(wrapper).toMatchSnapshot();
         });